==> Step_2_Creating_and_concatenating_cubes.ipynb <==
# For concatenation
import iris
from iris.util import equalise_attributes

# For plotting checks
import iris.quickplot as qplt
import matplotlib.pyplot as plt
import cartopy.crs as ccrs
import geopandas as gpd

## Concatenation

# Import data
tas1 = iris.load_cube('/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/tas/ERA5-Land_tas_hourly_1950_1954.nc')
tas2 = iris.load_cube('/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/tas/ERA5-Land_tas_hourly_1955_1959.nc')
tas3 = iris.load_cube('/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/tas/ERA5-Land_tas_hourly_1960_1964.nc')
tas4 = iris.load_cube('/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/tas/ERA5-Land_tas_hourly_1965_1969.nc')
tas5 = iris.load_cube('/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/tas/ERA5-Land_tas_hourly_1970_1974.nc')
tas6 = iris.load_cube('/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/tas/ERA5-Land_tas_hourly_1975_1979.nc')
tas7 = iris.load_cube('/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/tas/ERA5-Land_tas_hourly_1980_1984.nc')
tas8 = iris.load_cube('/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/tas/ERA5-Land_tas_hourly_1985_1989.nc')
tas9 = iris.load_cube('/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/tas/ERA5-Land_tas_hourly_1990_1994.nc')
tas10 = iris.load_cube('/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/tas/ERA5-Land_tas_hourly_1995_1999.nc')
tas11 = iris.load_cube('/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/tas/ERA5-Land_tas_hourly_2000_2004.nc')
tas12 = iris.load_cube('/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/tas/ERA5-Land_tas_hourly_2005_2009.nc')
tas13 = iris.load_cube('/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/tas/ERA5-Land_tas_hourly_2010_2014.nc')
tas14 = iris.load_cube('/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/tas/ERA5-Land_tas_hourly_2015_2019.nc')
tas15 = iris.load_cube('/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/tas/ERA5-Land_tas_hourly_2020_2024.nc')

# Create a list of cubes
tas_list = iris.cube.CubeList([tas1, tas2, tas3, tas4, tas5, tas6, tas7, tas8, tas9, tas10, tas11, tas12, tas13, tas14, tas15])

# Remove attributes which can cause concatenation issues
tas_removed_attributes = equalise_attributes(tas_list)
print(tas_removed_attributes)

# Concatenate the cubes
tas_cube_concatenated = tas_list.concatenate_cube()

iris.save(tas_cube_concatenated, '/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/ERA5-Land_tas_hourly_1950_2024.nc')

td1 = iris.load_cube('/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/td/ERA5-Land_td_hourly_1950_1954.nc')
td2 = iris.load_cube('/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/td/ERA5-Land_td_hourly_1955_1959.nc')
td3 = iris.load_cube('/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/td/ERA5-Land_td_hourly_1960_1964.nc')
td4 = iris.load_cube('/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/td/ERA5-Land_td_hourly_1965_1969.nc')
td5 = iris.load_cube('/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/td/ERA5-Land_td_hourly_1970_1974.nc')
td6 = iris.load_cube('/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/td/ERA5-Land_td_hourly_1975_1979.nc')
td7 = iris.load_cube('/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/td/ERA5-Land_td_hourly_1980_1984.nc')
td8 = iris.load_cube('/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/td/ERA5-Land_td_hourly_1985_1989.nc')
td9 = iris.load_cube('/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/td/ERA5-Land_td_hourly_1990_1994.nc')
td10 = iris.load_cube('/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/td/ERA5-Land_td_hourly_1995_1999.nc')
td11 = iris.load_cube('/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/td/ERA5-Land_td_hourly_2000_2004.nc')
td12 = iris.load_cube('/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/td/ERA5-Land_td_hourly_2005_2009.nc')
td13 = iris.load_cube('/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/td/ERA5-Land_td_hourly_2010_2014.nc')
td14 = iris.load_cube('/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/td/ERA5-Land_td_hourly_2015_2019.nc')
td15 = iris.load_cube('/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/td/ERA5-Land_td_hourly_2020_2024.nc')

# Create a list of cubes
td_list = iris.cube.CubeList([td1, td2, td3, td4, td5, td6, td7, td8, td9, td10, td11, td12, td13, td14, td15])

# Remove attributes which can cause concatenation issues
td_removed_attributes = equalise_attributes(td_list)
print(td_removed_attributes)

# Concatenate the cubes
td_cube_concatenated = td_list.concatenate_cube()

iris.save(td_cube_concatenated, '/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/ERA5-Land_td_hourly_1950_2024.nc')


## Masking using a shapefile

# Loading the data
tas_cube = iris.load_cube('/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/ERA5-Land_tas_hourly_1950_2024.nc')
td_cube = iris.load_cube('/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/ERA5-Land_td_hourly_1950_2024.nc')

# Loading and editing the shapefile to create a shapefile outline
south_africa = gpd.read_file('/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/Shapefiles/LocalMunicipalities2018_Final.shp')
limpopo = south_africa[south_africa['PROVINCE'] == 'LIM']
limpopo_outline = limpopo.dissolve()
limpopo_outline = limpopo_outline.to_crs(epsg=4326)
limpopo_outline.plot()

# Loading the data
    # tas_cube = iris.load_cube('/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/ERA5-Land_tas_hourly_1950_2024.nc')
    # td_cube = iris.load_cube('/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/ERA5-Land_td_hourly_1950_2024.nc')
    # limpopo_shapefile = gpd.read_file('/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Limpopo_Province_shapefile.geojson')

# Extracting the data

tas_cube_masked = iris.util.mask_cube_from_shapefile(tas_cube, limpopo_outline.geometry[0], minimum_weight=0.5)
tas_cube_masked.fill_value = -9999

td_cube_masked = iris.util.mask_cube_from_shapefile(td_cube, limpopo_outline.geometry[0], minimum_weight=0.5)
td_cube_masked_fill_value = -9999

# Saving the data
iris.save(tas_cube_masked, '/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/ERA5-Land_tas_hourly_1950_2024_Limpopo.nc')
iris.save(td_cube_masked, '/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/ERA5-Land_td_hourly_1950_2024_Limpopo.nc')



# Checking the masking minimum weight
masked_tas_cube = tas_cube_masked.copy().collapsed('time', iris.analysis.MEAN)

# Create a figure with the projection you want
fig, ax = plt.subplots(figsize=(10, 10), subplot_kw={'projection': ccrs.PlateCarree()})

qplt.pcolormesh(masked_tas_cube, cmap='Reds', axes=ax)

# Overlay the GeoDataFrame, ensuring it's in the same CRS
limpopo_outline.to_crs(epsg=4326).plot(ax=ax, facecolor='none', edgecolor='black')

plt.show()
