==> Step_3_Cube_creation_frontend.ipynb <==
Running this code will compare 21 heat index algorithms from Anderson et al (2013) to Lu and Romps (2022)
If using hourly-level data, expect considerable waiting times (8+ hours)

%run Step_3_Cube_creation_backend.ipynb

## Step 1: Specify filepaths and alike (see below for instructions)

# //NOTE: THIS CODE REQUIRES ERA5-LAND HOURLY DATA FOR tas and td

# tas: temperature at surface (in ˚C)
# rh: relative humidity (in %)
# td: dew point temperature (in ˚C)
# es: water vapor pressure (in kilopascals)

# Specify filepaths
    # If you have already run the Lu and Romps algorithm to create the 22n'd heat index cube, you can use a file path to that nc file to save a lot of computational time

filepaths = {'tas': '/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/ERA5-Land_tas_hourly_1950_2024_Limpopo.nc',
            'td': '/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/ERA5-Land/ERA5-Land_td_hourly_1950_2024_Limpopo.nc',
            # 'heat_index_22': '/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/heat_index_22.nc'
            }

# Specify start and end year of analyses
start_year = 1950
end_year = 2024

# Specify location name
location_name = 'Limpopo'

# Specify shapefile to work with
shapefile = gpd.read_file('/Users/maxwhite/Documents/Met_Office_Work/Heat_index_Algorithms/Data/Limpopo/Limpopo_Boundaries.geojson')

# Specify an output folder for figures
output_folder = f'{os.getcwd()}/{location_name}_outputs'
if not os.path.exists(output_folder):
    os.makedirs(output_folder)

unprocessed_output_folder = f'{output_folder}/basic_plots_for_unprocessed_data'
if not os.path.exists(unprocessed_output_folder):
    os.makedirs(unprocessed_output_folder)

## Step 2: Create the basic tas and td cubes

create_my_files() # creates tas, rh, td, and es cubes (so named, e.g. "tas_cube") unless they already exist

# Speedier version of loading the files if all goes well above but the kernel crashes later on

# def load_my_files():
    # tas_cube = iris.load_cube(f'{location_name}_tas_hourly_{start_year}-{end_year}.nc')
    # td_cube = iris.load_cube(f'{location_name}_td_hourly_{start_year}-{end_year}.nc')
    # es_cube = iris.load_cube(f'{location_name}_es_hourly_{start_year}-{end_year}.nc')
    # rh_cube = iris.load_cube(f'{location_name}_rh_hourly_{start_year}-{end_year}.nc')

## Step 3: Create the first 21 heat index cubes

create_all_21_heat_indices_cubes()

## Step 4: Create the Lu and Romps (2022) cube

heat_index_lu_and_romps_vectorized_pre_processing(tas_cube, rh_cube)

heat_index_lu_and_romps_vectorized_processing(daily_tas_cubes, daily_rh_cubes)

heat_index_lu_and_romps_vectorized_post_processing(final_heat_index_22_cube)
